src/uwsgi/requirements.txt
Flask
uwsgi
mysql.connector
bcrypt
Werkzeug
cachelib
pycryptodome